Move shared methods into abstract parent
