+++ inc/Engine/Optimization/GoogleFonts/AbstractGFOptimization.php
@@ -1,0 +1,84 @@
+<?php
+
+declare( strict_types=1 );
+
+namespace WP_Rocket\Engine\Optimization\GoogleFonts;
+
+use WP_Rocket\Engine\Optimization\AbstractOptimization;
+
+/**
+ * Abstract Optimization Parent Class for Google Fonts Optimizers.
+ *
+ * @since 3.8
+ */
+abstract class AbstractGFOptimization extends AbstractOptimization {
+
+	/**
+	 * Allowed display values.
+	 *
+	 * @since 3.8
+	 *
+	 * @var array
+	 */
+	protected $display_values = [
+		'auto'     => 1,
+		'block'    => 1,
+		'swap'     => 1,
+		'fallback' => 1,
+		'optional' => 1,
+	];
+
+	/**
+	 * Returns font with display value.
+	 *
+	 * @since  3.8 Moved here from GoogleFonts\Combine::class
+	 * @since  3.5.1
+	 *
+	 * @param array $font Array containing font tag and matches.
+	 *
+	 * @return string Google Font tag with display param.
+	 */
+	protected function get_font_with_display( array $font ) {
+		$font_url = html_entity_decode( $font['url'] );
+		$query    = wp_parse_url( $font_url, PHP_URL_QUERY );
+
+		if ( empty( $query ) ) {
+			return $font[0];
+		}
+
+		$display     = $this->get_font_display_value();
+		$parsed_font = wp_parse_args( $query );
+
+		$font_url = ! empty( $parsed_font['display'] )
+			? str_replace( "&display={$parsed_font['display']}", "&display={$display}", $font_url )
+			: "{$font_url}&display={$display}";
+
+		return str_replace( $font['url'], esc_url( $font_url ), $font[0] );
+	}
+
+	/**
+	 * Get the font display value.
+	 *
+	 * @since  3.8 Moved here from GoogleFonts\Combine::class
+	 * @since  3.5.1
+	 *
+	 * @return string font display value.
+	 */
+	protected function get_font_display_value(): string {
+		/**
+		 * Filters the combined Google Fonts display parameter value
+		 *
+		 * @since  3.8 Moved here from GoogleFonts\Combine::class
+		 * @since  3.3.5
+		 *
+		 * @param string $display Display value. Can be either auto, block, swap, fallback or optional.
+		 */
+		$display = apply_filters( 'rocket_combined_google_fonts_display', 'swap' );
+
+		if ( ! is_string( $display ) ) {
+			return 'swap';
+		}
+
+		return isset( $this->display_values[ $display ] ) ? $display : 'swap';
+	}
+}
